frontend Functionality Update

+++ src/app/groups/groups.component.ts
@@ -1,8 +1,8 @@
-import { Component , HostListener } from '@angular/core';
+import { Component } from '@angular/core';
 import {NavigationExtras, Router} from '@angular/router';
 import { GroupDataService } from '../group-data.service';
 import {DeleteGroupService} from "../delete-group.service";
-
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -29,7 +29,7 @@ export class GroupsComponent {
   }
 
 
-  headers = ["ID", "Group Name", "Actions"];
+  headers = ["Group Name", "Actions"];
 
   
 
@@ -39,22 +39,48 @@ export class GroupsComponent {
     alert( id)
     let navigationExtras: NavigationExtras = {
       queryParams: {
-        groupID: id
+        groupID: id,
+        
       }
     };
     this.router.navigate(['/group/create'],navigationExtras)
   }
 
-  performView(id: number){
-    this.router.navigate(['/groups/'+id])
+  performView(id: number, name: string){
+    //this.router.navigate(['/groups/'+id])
+    let navigationExtras: NavigationExtras = {
+      queryParams: {
+        groupID: id,
+        groupName: name
+        
+      }
+    };
+    this.router.navigate(['/groups/'+id],navigationExtras)
   }
+  
 
   performDelete(id: number,name:String){
-    alert(name)
-    this.del.deleteData(id)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Are you sure, you want to delete this group ?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, go ahead.',
+      cancelButtonText: 'No, let me think',
+    }).then((result) => {
+      if (result.value) {
+        this.del.deleteData(id)
+        Swal.fire('Removed!', 'Group successfully deleted', 'success').then(()=>{
+          location.reload();
+        });
+
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire('Cancelled', 'Group is still in the database', 'error');
+      }
+    });
   }
 
-
+  
   performCreate() {
     this.router.navigate(['/group/create'])
   }
